chore: css fixes - move contribute-menu to sub folder under nav
